+++ lib/glimmer/libui/custom_control/code_entry.rb
@@ -78,10 +78,12 @@ module Glimmer
           text(padding, padding) {
             default_font @font_default
             
+            # TODO cache this work if not changed between renders
             syntax_highlighter.syntax_highlighting(code).each do |token|
               style_data = Rouge::Theme.find(theme).new.style_for(token[:token_type])
-             
-              string(token[:token_text]) {
+              token_text = token[:token_text].start_with?("\n") ? " #{token[:token_text]}" : token[:token_text]
+              
+              string(token_text) {
                 color style_data[:fg] || :black
                 background style_data[:bg] || :white
               }
@@ -96,7 +98,7 @@ module Glimmer
             # TODO perhaps instead of this approach, try to calculate text extent as that is supported by libui
             string(caret_text) {
               color :black
-              background :white
+              background [0, 0, 0, 0]
             }
           }
         end